FAIL2.0

+++ src/Network/Service.java
@@ -13,13 +13,13 @@ import java.math.BigDecimal;
  * @author Rui Barcelos https://github.com/barcelosrui
  */
 public class Service implements Serializable {
-    private double cx;
-    private double cy;
+    private BigDecimal cx;
+    private BigDecimal cy;
     private BigDecimal zoom;
     private byte[] data;
             int imageNumber;
     
-    public Service(double cx, double cy, BigDecimal zoom){
+    public Service(BigDecimal cx, BigDecimal cy, BigDecimal zoom){
         this.cx = cx;
         this.cy = cy;
         this.zoom = zoom;
@@ -35,19 +35,19 @@ public class Service implements Serializable {
         return s;
     }
     
-    public double getCx() {
+    public BigDecimal getCx() {
         return cx;
     }
 
-    public void setCx(double cx) {
+    public void setCx(BigDecimal cx) {
         this.cx = cx;
     }
 
-    public double getCy() {
+    public BigDecimal getCy() {
         return cy;
     }
 
-    public void setCy(double cy) {
+    public void setCy(BigDecimal cy) {
         this.cy = cy;
     }
